Some more minor cleanup, also moved the TrajectoryGenerator to the example.
